improve code based on PulRqst comments
