+++ src/pages/Product/InfoPanel/InfoPanel.module.scss
@@ -30,6 +30,3 @@
 .tab {
   padding: 10px 20px;
 }
-.comments__container {
-  margin: 30px 90px 0 90px;
-}
