+++ src/pages/Product/InfoPanel/Reviews/CommentForms/MyTextArea.js
@@ -3,10 +3,10 @@ import React from 'react';
 import styles from './MyTextArea.module.scss';
 
 const MyTextArea = ({ field, form, ...rest }) => (
-  <div className={styles.center}>
+  <div>
     <textarea {...field} {...rest} />
     {form.errors[field.name] && form.touched[field.name] && (
-      <span className={styles.error}>{form.errors[field.name]}</span>
+      <div className={styles.error}>{form.errors[field.name]}</div>
     )}
   </div>
 );
