+++ src/pages/Product/InfoPanel/Reviews/CommentForms/MyTextArea.module.scss
@@ -1,0 +1,5 @@
+.error {
+  color: red;
+  margin-bottom: 20px;
+}
+
