+++ src/pages/Product/InfoPanel/Reviews/CommentsFilter/CommentsFilter.module.scss
@@ -1,6 +1,7 @@
-@import '../../../../utils/styles/var';
+@import '../../../../../utils/styles/var';
 .filter {
   text-align: center;
+  margin-top: 25px;
 }
 .text {
   font-size: 14px;
