--- src/pages/Product/InfoPanel/Reviews/MyTextArea.module.scss
@@ -1,8 +1,0 @@
-.error {
-  color: red;
-  margin-left: 25px;
-}
-.center {
-  display: flex;
-  align-items: center;
-}
