+++ src/pages/Product/InfoPanel/Reviews/Reviews.module.scss
@@ -1,0 +1,6 @@
+.comments__container {
+    margin: 30px 90px 0 90px;
+  }
+.message {
+    color:red;
+}
